✨ ajuste en los componentes
