+++ src/app/components/articles/article/article.component.html
@@ -1,17 +1,12 @@
-<section class="post">
-  <a routerLink="/blog" *ngFor="let post of posts">
-    <article class="post__card">
-      <div class="post__card-image">
-        <img [src]="post.cover_image" [alt]="post.title">
-      </div>
-      <div class="post__card-resumen">
-        <h1>{{post.title}}</h1>
-        <span>{{post.published_at | date}}</span>
-        <p>{{post.description}}</p>
-      </div>
-    </article>
-  </a>
-  <div class="post__redirect">
-    <a routerLink="/blog">Ver todos los posts -></a>
-  </div>
-</section>
+<a routerLink="/blog">
+  <article class="card">
+    <div class="card__image">
+      <img [src]="posts.cover_image" [alt]="posts.title">
+    </div>
+    <div class="card__resumen">
+      <h1>{{posts.title}}</h1>
+      <span>{{posts.published_at | date}}</span>
+      <p>{{posts.description}}</p>
+    </div>
+  </article>
+</a>
