+++ src/app/components/articles/article/article.component.scss
@@ -1,44 +1,61 @@
-.post {
+.card {
     display: flex;
-    flex-direction: column;
+    flex-direction: row;
+    justify-content: space-between;
+    width: 100%;
+    min-height: 132px;
+    height: auto;
+    gap: 1em;
 
-    a {
-        margin: 3px 0;
-    }
-
-    &__card {
-        display: flex;
-        flex-direction: row;
-        justify-content: space-between;
-        width: 100%;
+    &__image {
+        width: 320px;
         height: 132px;
-        gap: 1em;
 
-        &-image {
+        img {
             width: 320px;
-            height: 132px;
-
-            img {
-                width: 320px;
-                height: 100%;
-                border-radius: 5px;
-            }
-        }
-
-        &-resumen {
-            display: flex;
-            flex-direction: column;
-            justify-content: space-around;
-
-            p {
-                margin: 0;
-                padding: 0;
-            }
-
-            h1 {
-                margin: 0;
-                padding: 0;
-            }
+            height: 100%;
+            border-radius: 5px;
         }
     }
+
+    &__resumen {
+        display: flex;
+        flex-direction: column;
+        justify-content: space-around;
+
+        p {
+            margin: 0;
+            padding: 0;
+        }
+
+        h1 {
+            margin: 0;
+            padding: 0;
+        }
+    }
+
+    // &__redirect {
+    //     display: flex;
+    //     flex-direction: row;
+    //     justify-content: center;
+    //     align-items: center;
+    //     gap: 1rem;
+    // }
+}
+
+
+button {
+    margin: 0;
+    padding: 0;
+    width: 50%;
+    height: 46px;
+    border-radius: 5px;
+    background-color: #000;
+    color: #fff;
+    text-align: center;
+    font-size: 1.2rem;
+    font-weight: bold;
+    text-decoration: none;
+    transition: all 0.3s ease-in-out;
+    cursor: pointer;
 }
