+++ src/app/components/banner/banner/banner.component.html
@@ -6,9 +6,9 @@
     <article class="banner__technologies-card">
       <span>🖖</span>
       <div>
-        <p>Hola, yo soy</p>
+        <p>Hola, soy</p>
         <h1>Danniel Navas</h1>
-        <p>Frontend developer.</p>
+        <p>JavaScript developer.</p>
       </div>
     </article>
     <div class="banner__technologies-icons">
