+++ src/app/pages/portal/principal/principal.component.ts
@@ -1,4 +1,5 @@
 import { Component, OnInit } from '@angular/core';
+import { IPost } from '@core/models/devto/response-devto';
 import { IResponseStatistics } from '@core/models/statistic/response-statistec';
 import { RequestYoutube } from '@core/models/youtube/request-youtube';
 import { ResponseYoutube } from '@core/models/youtube/response-youtube';
@@ -13,9 +14,14 @@ import { YoutubeService } from '@core/services/youtube/youtube.service';
 export class PrincipalComponent implements OnInit {
   datos = [1, 2, 3];
   dataYoutube!: ResponseYoutube;
-  constructor(private youtubeService: YoutubeService) {}
+  posts!: IPost[];
+  constructor(
+    private youtubeService: YoutubeService,
+    private devtoService: DevtoService
+  ) {}
   ngOnInit(): void {
     this.getDataYoutube();
+    this.getPostsDevto();
   }
 
   getDataYoutube(): void {
@@ -24,5 +30,10 @@ export class PrincipalComponent implements OnInit {
       .subscribe((data: ResponseYoutube) => {
         this.dataYoutube = data;
       });
+  }
+  getPostsDevto(): void {
+    this.devtoService.getPostDevTo().subscribe((data: IPost[]) => {
+      this.posts = data.slice(0, 3);
+    });
   }
 }
