+++ src/app/shared/pipes/repo-list-pipe.pipe.ts
@@ -6,22 +6,25 @@ import { IResponseGithub } from '@core/models/github/response-github';
 })
 export class RepoListPipe implements PipeTransform {
   transform(value: any[], ...args: number[]): any[] {
-    let repos = [];
-    if (Number(args[0]) !== 0) {
-      let random = 0;
-      let history = [];
-      for (let i = 0; i < args[0]; i++) {
-        random = Math.floor(Math.random() * value.length);
-        if (history.filter((item) => item === random).length === 0) {
-          repos.push(value[random]);
-          history.push(random);
-        } else {
-          i--;
+    if (value) {
+      let repos = [];
+      if (Number(args[0]) !== 0) {
+        let random = 0;
+        let history = [];
+        for (let i = 0; i < args[0]; i++) {
+          random = Math.floor(Math.random() * value.length);
+          if (history.filter((item) => item === random).length === 0) {
+            repos.push(value[random]);
+            history.push(random);
+          } else {
+            i--;
+          }
         }
+      } else {
+        repos = value;
       }
-    } else {
-      repos = value;
+      return repos;
     }
-    return repos;
+    return [];
   }
 }
